# src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Wine Recommendation API
  version: 1.0.0
  description: API for wine recommendations based on user preferences

servers:
  - url: http://localhost:3000/api
    description: Development server

paths:
  /recommendations:
    post:
      summary: Get wine recommendations
      description: Returns personalized wine recommendations based on user preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
      responses:
        '200':
          description: Successful recommendation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        '400':
          description: Invalid input
        '500':
          description: Server error

  /search:
    get:
      summary: Search wines
      description: Search wines by query parameters
      parameters:
        - $ref: '#/components/parameters/queryParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/regionParam'
        - $ref: '#/components/parameters/minPriceParam'
        - $ref: '#/components/parameters/maxPriceParam'
        - $ref: '#/components/parameters/pageParam'
      responses:
        '200':
          description: Successful search response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wine'
        '500':
          description: Server error

components:
  schemas:
    RecommendationRequest:
      type: object
      properties:
        userId:
          type: string
        preferences:
          type: object
          properties:
            wineType:
              type: string
            priceRange:
              type: array
              items:
                type: number
      required:
        - userId
        - preferences

    Wine:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        region:
          type: string
        price:
          type: number
        rating:
          type: number

  parameters:
    queryParam:
      name: query
      in: query
      description: Search query
      required: true
      schema:
        type: string

    limitParam:
      name: limit
      in: query
      description: Maximum number of results
      schema:
        type: integer
        default: 10

    offsetParam:
      name: offset
      in: query
      description: Number of results to skip
      required: false
      schema:
        type: integer
        default: 0

    regionParam:
      name: region
      in: query
      description: Filter by region
      required: false
      schema:
        type: string

    minPriceParam:
      name: minPrice
      in: query
      description: Minimum price
      required: false
      schema:
        type: number

    maxPriceParam:
      name: maxPrice
      in: query
      description: Maximum price
      required: false
      schema:
        type: number

    pageParam:
      name: page
      in: query
      description: Page number for pagination (overrides offset)
      required: false
      schema:
        type: integer
        default: 1